Add support for PHC error bound interface
With EC2 introducing a PTP Hardware Clock on instances, Chrony
under-reports the clock error bound, assuming that it is 0 when that is
not the case. The ENA driver exposes an interface for gathering the
clock error bound of the PTP Hardware Clock (PHC) itself. We need to add
this interface's error bound value to the clock error bound reported
from ClockBound so that we do not underestimate our clock error.

This commit adds support for users to supply a PHC reference ID (this
comes from chronyd, generally will be PHC0) and and a PHC interface (the
network interface for ENA driver where PHC is exposed). If supplied both
of these arguments, Clockbound will, whenever it polls Chrony for error
bound information, add the PHC error bound to its own error bound
reported if Chrony is syncing to the PHC (otherwise 0).