Update PHC error bound feature failure cases
In our first iteration of support for PHC error bound interface in
ClockBound, we made it possible to panic upon failure to read the PHC
error bound sysfs file, which could cause unexpected crashes in the
ClockBound daemon. Now, we handle such errors when reading the sysfs
file.

Additionally, we migrate some of the initialization logic for
discovering the path of the sysfs error bound file into the main entry
point instead and allow PhcInfo to hold that path instead of the
interface used to discover that path. To have finer grained control over
the output in stderr when `main` runs into an Err case, we also change
typing to use a String instead of printing std::io::Error contents.